V1.1.0

+++ backend/backend/DTOs/Request/Parcel/CUParcelRequest.cs
@@ -6,5 +6,5 @@ public class CUParcelRequest
 {
     [Required] public string Name { get; set; }
     [Required] public long Number { get; set; }
-    [Required] public long Size { get; set; }
+    [Required] public double Size { get; set; }
 }
